+++ fabric-object-builder-api-v1-bridge/src/main/java/dev/loaderbridge/fabric/api/object/builder/FabricObjectBuilderBridgeProvider.java
@@ -15,12 +15,14 @@ public final class FabricObjectBuilderBridgeProvider implements RuntimeBridgeMod
     private static final RuntimeBridgeModule DESCRIPTOR = new RuntimeBridgeModule(
             "fabric-object-builder-api-v1-bridge",
             "fabric-object-builder-api-v1:15.2.1",
-            "15.2.1+40875a9319-loaderbridge.6",
+            "15.2.1+40875a9319-loaderbridge.7",
             BridgeCapability.FABRIC_API,
             Set.of(
                     "net.fabricmc.fabric.api.object.builder.v1.block.entity.FabricBlockEntityTypeBuilder",
                     "net.fabricmc.fabric.api.object.builder.v1.block.entity."
                             + "FabricBlockEntityTypeBuilder$Factory",
+                    "net.fabricmc.fabric.api.object.builder.v1.block.entity.FabricBlockEntityType",
+                    "net.fabricmc.fabric.api.object.builder.v1.block.entity.FabricBlockEntityType$Builder",
                     "net.fabricmc.fabric.api.object.builder.v1.entity."
                             + "FabricDefaultAttributeRegistry",
                     "net.fabricmc.fabric.api.object.builder.v1.entity.FabricEntityTypeBuilder",